+++ pygopherd/protocols/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ["base", "enhanced", "gemini", "gopherp", "rfc1436", "http", "wap"]
+__all__ = ["base", "enhanced", "gemini", "gopherp", "rfc1436", "http", "spartan", "wap"]